+++ nf_core/modules/lint/module_deprecations.py
@@ -8,12 +8,12 @@ def module_deprecations(_, module):
     """
     Check that the modules are up to the latest nf-core standard
     """
-    module.wf_path = module.module_dir
-    if "functions.nf" in os.listdir(module.module_dir):
+    module.wf_path = module.component_dir
+    if "functions.nf" in os.listdir(module.component_dir):
         module.failed.append(
             (
                 "module_deprecations",
                 "Deprecated file `functions.nf` found. No longer required for the latest nf-core/modules syntax!",
-                module.module_dir,
+                module.component_dir,
             )
         )