+++ nf_core/modules/modules_utils.py
@@ -3,7 +3,7 @@ import os
 import urllib
 from pathlib import Path
 
-from .nfcore_module import NFCoreModule
+from ..components.nfcore_component import NFCoreComponent
 
 log = logging.getLogger(__name__)
 
@@ -81,10 +81,17 @@ def get_installed_modules(dir, repo_type="modules"):
             else:
                 nfcore_modules.append(m)
 
-    # Make full (relative) file paths and create NFCoreModule objects
+    # Make full (relative) file paths and create NFCoreComponent objects
     local_modules = [os.path.join(local_modules_dir, m) for m in local_modules]
     nfcore_modules = [
-        NFCoreModule(m, "nf-core/modules", Path(nfcore_modules_dir, m), repo_type=repo_type, base_dir=Path(dir))
+        NFCoreComponent(
+            m,
+            "nf-core/modules",
+            Path(nfcore_modules_dir, m),
+            repo_type=repo_type,
+            base_dir=Path(dir),
+            component_type="modules",
+        )
         for m in nfcore_modules
     ]