+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### Linting
 
+- Add new command `nf-core subworkflows lint` ([#2379](https://github.com/nf-core/tools/pull/2379))
+
 ### Modules
 
 ### Subworkflows